[y][lib/yrenderer] update

+++ src/lib/yrenderer/post/MultisampleResolver.cpp
@@ -25,8 +25,8 @@ MultisampleResolver::MultisampleResolver(Context* ctx, ygfx::Texture* tex_ms, yg
 void MultisampleResolver::render(const RenderParams& params) {
 	// resolve
 	if (true) {
-		tsr->bindings.shader_data.dict_set("width:0", into_texture->frame_buffer->width);
-		tsr->bindings.shader_data.dict_set("height:4", into_texture->frame_buffer->height);
+		tsr->bindings.shader_data.dict_set("width:0", (float)into_texture->frame_buffer->width);
+		tsr->bindings.shader_data.dict_set("height:4", (float)into_texture->frame_buffer->height);
 		tsr->set_source(dynamicly_scaled_source());
 		into_texture->set_area(dynamicly_scaled_area(into_texture->frame_buffer.get()));
 		into_texture->render(params);
